updated return data type

+++ libs/ProxomitySensorController.h
@@ -5,7 +5,7 @@ class ProximitySensorController
 {
 public:
     ProximitySensorController(uint8_t trig_pin, uint8_t echo_pin, uint16_t min_offset, uint16_t max_offset);
-    uint8_t measure(uint8_t measurement_time);
+    uint16_t measure(uint8_t measurement_time);
     void setUp();
     bool compare(uint16_t mesured_len);
     uint8_t t_pin;
